# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CalculEat API Documentation - OpenAPI 3.0",
    "description": "Sample description.",
    "version": "1.0.10"
  },
  "servers": [
    {
      "url": "http://hostname:7854/"
    }
  ],
...
  "paths": {
...
    "/api/1.0/predict": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "AI prediction endpoint",
        "description": "Calculate macros of given food image",
        "operationId": "predict",
        "requestBody": {
          "description": "Calculate macros of given food image",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponseOK"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponseBadRequest"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Authentication failed"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "err": {
                      "type": "string",
                      "example": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/1.0/message": {
      "get": {
        "tags": [
          "message"
        ],
        "summary": "List all of the messages in DB.",
        "description": "Fetch all of the messages from DB.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The id of sender that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 12
            }
          },
          {
            "name": "reciever_id",
            "in": "query",
            "description": "The id of sender that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 256
            }
          },
          {
            "name": "creation_time",
            "in": "query",
            "description": "The creation datetime of message that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2017-07-21T17:32:28Z"
            }
          },
          {
            "name": "message_text",
            "in": "query",
            "description": "The message text that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Hello world"
            }
          }
        ],
        "operationId": "listMessages",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponseBody"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:messages",
              "read:messages"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "message"
        ],
        "summary": "Add a new message to the database",
        "description": "Add a new Message to the database. All of the fields are not needed to create meeting.",
        "operationId": "createMessage",
        "requestBody": {
          "description": "Create a new message in the DB.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:meetings",
              "read:meetings"
            ]
          }
        ]
      }
    },
    "/api/1.0/message/{message_id}": {
      "patch": {
        "tags": [
          "message"
        ],
        "summary": "Update message",
        "description": "This can only be done by the logged in user. **You can use only fields that you want to update.**",
        "operationId": "updateMessage",
        "requestBody": {
          "description": "Update the message object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequestBody"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "description": "The message_id that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePATCHResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request"
          },
          "404": {
            "description": "Meeting not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:meetings",
              "read:meetings"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "message"
        ],
        "summary": "Deletes the message from the database. This can only be done by the logged in user.",
        "description": "Deletes the specified message according to the ID.",
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "description": "The message_id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message successfuly deleted."
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Message not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:messages",
              "read:messages"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessagePATCHResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "sender_id": {
                "type": "integer",
                "format": "int64",
                "example": 11
              },
              "reciever_id": {
                "type": "integer",
                "format": "int64",
                "example": 11
              },
              "message_text": {
                "type": "string",
                "example": "Hello, I'm a client!"
              },
              "CreatedAt": {
                "type": "string",
                "example": "2023-09-10T18:25:55.6236225+03:00"
              },
              "UpdatedAt": {
                "type": "string",
                "example": "2023-09-10T18:25:55.6236225+03:00"
              },
              "deleted_on": {
                "type": "string",
                "example": "0001-01-01T00:00:00Z"
              }
            }
          },
          "msg": {
            "type": "string",
            "example": "Successful operation, object has been updated."
          }
        }
      },
      "MessageRequestBody": {
        "type": "object",
        "properties": {
          "reciever_id": {
            "type": "integer",
            "format": "int64",
            "example": 11
          },
          "message_text": {
            "type": "string",
            "example": "Hello, I'm a client!"
          }
        }
      },
      "MessageResponseBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "sender_id": {
            "type": "integer",
            "format": "int64",
            "example": 11
          },
          "reciever_id": {
            "type": "integer",
            "format": "int64",
            "example": 11
          },
          "message_text": {
            "type": "string",
            "example": "Hello, I'm a client!"
          },
          "CreatedAt": {
            "type": "string",
            "example": "2023-09-10T18:25:55.6236225+03:00"
          },
          "UpdatedAt": {
            "type": "string",
            "example": "2023-09-10T18:25:55.6236225+03:00"
          },
          "deleted_on": {
            "type": "string",
            "example": "0001-01-01T00:00:00Z"
          }
        }
      },
...
      "PredictResponseOK": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "calories": {
                "type": "number",
                "example": 210.43324491600353
              },
              "protein": {
                "type": "number",
                "example": 4.370536625178535
              },
              "fat": {
                "type": "number",
                "example": 0.8093586342923212
              },
              "carbs": {
                "type": "number",
                "example": 45.32408352036999
              }
            }
          },
          "msg": {
            "type": "string",
            "example": "Successfully calculeated"
          }
        }
      },
      "PredictResponseBadRequest": {
        "type": "object",
        "properties": {
          "err": {
            "type": "string",
            "example": "Invalid Request Format, couldnt bind predict object's json"
          }
        }
      },
      "PredictRequestBody": {
        "type": "object",
        "properties": {
          "file": {
            "type": "object",
            "format": "form-data",
            "example": "image.png"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}